gameState.Enemy implemented & etc
Enemy can overlap with player which results in losing the game.
This also pauses the game and all animations.
Text will be shown and game can be restart.

Player will move by itself to reach the goal.
Player has abilities to jump over obstacles which is later introduced.
When it reaches the repel radius * 1/3, then it will slows down as if to rest and overlaps with the goal.