New obstacles and updated path generation
Paths with be generated with the increment of score.

Edited the scrollfactor of the background to match the paths.

New obstacles introduces, which is the tree that block the path of the player, and it will randomly generated on each platform.

2 more paths was introduced so player could switch the advancement path.

Added a pausing menu at the start of the game, so that player and enemy could initialize correctly

Changed the goal position to match the game score.